third_party/WebKit/LayoutTests/editing/style/bold_italic_linebreak.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      'here is <i>^some</i> text<br>',
      'here| is some more text',
    '</div>'
  ].join(''),
  'bold',
  [
    '<div contenteditable>',
      'here is <b><i>^some</i> text<br>',
      'here|</b> is some more text',
    '</div>'
  ].join('')), 'Bold with itatic and line break');
</script>
